requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-bootstrap3==11.0.0
django-braces==1.13.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.9.2
djoser==2.0.3
fcm-django==0.3.4
gunicorn==20.0.4
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
psycopg2-binary==2.8.2
pyfcm==1.4.7
PyJWT==1.7.1
pynetbox==4.0.6
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
stripe==2.27.0
urllib3==1.24.2
